requirements.txt
selenium~=4.0.0
requests~=2.26.0
numpy~=1.22.3
scikit-learn~=0.24.2
nltk~=3.6.7
tensorflow-gpu~=2.8.0
sqlparse~=0.4.2
Pillow~=9.0.0
SQLAlchemy~=1.4.32
psycopg2~=2.9.1
Flask~=1.1.1
gunicorn